UN-2773 Refactor run_context.langchain for better grouping
