+++ MANIFEST.in
@@ -14,5 +14,5 @@ include neuro_san/deploy/Dockerfile
 include neuro_san/deploy/entrypoint.sh
 include neuro_san/deploy/logging.json
 include neuro_san/deploy/run.sh
-include neuro_san/internals/run_context/langchain/default_llm_info.hocon
-include neuro_san/internals/run_context/langchain/toolbox_info.hocon
+include neuro_san/internals/run_context/langchain/llms/default_llm_info.hocon
+include neuro_san/internals/run_context/langchain/toolbox/toolbox_info.hocon
