+++ docs/llm_info_hocon_reference.md
@@ -46,7 +46,7 @@ Sub-keys to those dictionaries will be described in the next-level down heading 
 All parameters listed here have global scope and are listed at the top of the file by convention.
 
 The default file used with the system is called
-[default_llm_info.hocon](../neuro_san/internals/run_context/langchain/default_llm_info.hocon).
+[default_llm_info.hocon](../neuro_san/internals/run_context/langchain/llms/default_llm_info.hocon).
 
 ### Model Name Keys
 
@@ -225,7 +225,7 @@ look like this:
 Any classes listed must:
 
 * Exist in the PYTHONPATH of your server
-* Derive from neuro_san.internals.run_context.langchain.langchain_llm_factory.LangChainLlmFactory
+* Derive from neuro_san.internals.run_context.langchain.llms.langchain_llm_factory.LangChainLlmFactory
   to override the create_base_chat_model() method that creates your BaseLanguageModel instance.
 * Have a no-args constructor
