+++ neuro_san/internals/run_context/factory/master_llm_factory.py
@@ -13,7 +13,7 @@ from typing import Any
 from typing import Dict
 
 from neuro_san.internals.interfaces.context_type_llm_factory import ContextTypeLlmFactory
-from neuro_san.internals.run_context.langchain.default_llm_factory import DefaultLlmFactory
+from neuro_san.internals.run_context.langchain.llms.default_llm_factory import DefaultLlmFactory
 
 
 class MasterLlmFactory:
