+++ neuro_san/internals/run_context/factory/run_context_factory.py
@@ -16,7 +16,7 @@ from neuro_san.internals.interfaces.invocation_context import InvocationContext
 from neuro_san.internals.run_context.factory.master_llm_factory import MasterLlmFactory
 from neuro_san.internals.run_context.interfaces.run_context import RunContext
 from neuro_san.internals.run_context.interfaces.tool_caller import ToolCaller
-from neuro_san.internals.run_context.langchain.langchain_run_context import LangChainRunContext
+from neuro_san.internals.run_context.langchain.core.langchain_run_context import LangChainRunContext
 from neuro_san.internals.run_context.openai.openai_run_context import OpenAIRunContext
