+++ neuro_san/internals/run_context/langchain/llms/default_llm_info.hocon
@@ -554,7 +554,7 @@
         #
         # Any classes listed must:
         #   * Exist in the PYTHONPATH of your server
-        #   * Derive from neuro_san.internals.run_context.langchain.langchain_llm_factory.LangChainLlmFactory
+        #   * Derive from neuro_san.internals.run_context.langchain.llms.langchain_llm_factory.LangChainLlmFactory
         #   * Have a no-args constructor
 
         "openai": {
