+++ tests/neuro_san/internals/run_context/llms/test_llm_factory.py
@@ -20,7 +20,7 @@ from langchain_core.callbacks.base import BaseCallbackHandler
 from langchain_core.language_models.base import BaseLanguageModel
 from langchain_openai.chat_models.base import ChatOpenAI
 
-from neuro_san.internals.run_context.langchain.langchain_llm_factory import LangChainLlmFactory
+from neuro_san.internals.run_context.langchain.llms.langchain_llm_factory import LangChainLlmFactory
 
 
 class TestLlmFactory(LangChainLlmFactory):
